modified Import Excel to support intertemporal inputs
changed MainView to be able to select mutiple files of an intertemporal input.
This needed change of filename handling from a single string to a list of strings in Controller.
savetojson only needed to not pack input_files into a list as even single files will now be sent as list of a single string.

it is to be clarified whether selecting a folder for intertemporal inputs is more useful