added multiple try excepts to compensate for older excel files
read_commodities: removed multiple .item() as they throw errors.
sometimes int or float wrappers were used, these should be reviewed for necessity.
read_transmission: replaced if/else statement to check for data source row with try except.

in general try except should be preferable, but the position has to be adjusted, as an exorbitant amount of error infos is printed repeating the same information.
A possible solution would be a final error report.