creted global check for source
replaced all individual checks for the word "Source" with one global test. This raises a UserWarning and exits as individual removel would be time consuming. To-Do: Remove rows and columns with Source automatically.

Added test for "maxperhour". Older files have "maxperstep" instead, breaking conversion. If theres no maxperhour a Userwarning is raised.

Changed check of input filepath to more standard approach isinstance().

Added handling when  "Curtailment" exists in Process-Commodities. This is not done and needs fixing as it is possibly related to a bug where a Pandas Series is sent to json conversion. it is not clear how these are related but the series in question matches "Curtailment".

Added other filepaths for easier selection.